requirements.txt
aliyun-python-sdk-core-v3==2.9.0
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
beautifulsoup4==4.6.0
billiard==3.5.0.4
celery==4.2.1
celery-redbeat==0.11.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.2.10
django-autocomplete-light==3.2.10
django-braces==1.14.0
django-dbsettings==1.0.1
django-environ==0.4.4
django-extensions==1.9.7
django-filter==2.2.0
django-guardian==1.4.9
django-js-reverse==0.8.1
django-material==1.1.1
django-modelcluster==3.1
django-nose==1.4.5
django-polymorphic==2.0.2
django-redis==4.8.0
django-rest-swagger==2.2.0
django-stdimage==2.4.2
django-taggit==1.2.0
django-tenant-schemas==1.9.0
django-tinymce4-lite==1.6.0
django-treebeard==4.3
django-webtest==1.7.8
djangorestframework==3.11.0
djangorestframework-csv==1.3.0
drf-extensions==0.3.1
easy-thumbnails==2.5
gunicorn==18.0
html5lib==0.999999999
idna==2.6
itypes==1.1.0
Jinja2==2.10
jsmin==2.2.2
jsonfield==2.0.2
kombu==4.2.1
lxml==4.2.3
MarkupSafe==1.0
mock==1.0.1
mysqlclient==1.3.13
nose==1.3.7
numpy==1.13.3
olefile==0.45.1
openapi-codec==1.3.2
pathlib==1.0.1
Pillow==4.0.0
progressbar2==3.38.0
psycopg2-binary==2.7.5
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.6
pycurl==7.43.0.1  --global-option="--with-nss"
pyenchant==2.0.0
pyOpenSSL==17.3.0
PyPDF2==1.26.0
pypinyin==0.27.0
python-dateutil==2.7.3
python-doc-inherit==0.3.0
python-forex-quotes==1.0
python-http-client==3.1.0
python-utils==2.3.0
pytz==2017.3
redis==2.10.6
reportlab==3.5.5
requests==2.21.0
requests-file==1.4.3
simplejson==3.16.0
six==1.10.0
git+https://github.com/Telstra/MessagingAPI-SDK-python.git
tldextract==2.2.0
tqdm==4.24.0
typing==3.6.4
Unidecode==1.0.22
uritemplate==3.0.0
urlfetch==1.0.2
urllib3==1.22
uWSGI==2.0.12
vine==1.1.4
waitress==1.1.0
Wand==0.4.3
webencodings==0.5.1
WebOb==1.8.2
WebTest==2.0.30
wechat-sdk==0.6.4
weixin-python==0.5.4
Willow==1.0
xhtml2pdf==0.0.6
xmltodict==0.9.2
python-telegram-bot==11.1.0
